refactor: 보낸 메시지 자동 읽음 처리 서비스를 네이밍을 명확하게 하고 메서드 명도 변경
